Fix jom use
Jom stopped working and would fail with the following message:
"jom: Can't connect to job server: Cannot determine jobserver name."

It seems that chocolatey copy/pastes jom using shimgen: https://github.com/chocolatey/shimgen
a closed-source project and that it breaks jom.
However, using the original jom that chocolatey installs works fine. So we
do that instead. To get the path to the original jom, we use the environment
variable ChocolateyInstall that chocolatey sets when installed.